Update OpenMP in dcft

+++ psi4/src/psi4/dcft/dcft_density_UHF.cc
@@ -100,7 +100,7 @@ DCFTSolver::compute_unrelaxed_density_OOOO() {
         global_dpd_->buf4_mat_irrep_rd(&Gaa, h);
 
         #pragma omp parallel for
-        for(int ij = 0; ij < Gaa.params->rowtot[h]; ++ij){
+        for (long ij = 0; ij < Gaa.params->rowtot[h]; ++ij){
             size_t i = Gaa.params->roworb[h][ij][0];
             int Gi = Gaa.params->psym[i];
             i -= Gaa.params->poff[Gi];
@@ -144,7 +144,7 @@ DCFTSolver::compute_unrelaxed_density_OOOO() {
         global_dpd_->buf4_mat_irrep_rd(&Gab, h);
 
         #pragma omp parallel for
-        for(int ij = 0; ij < Gab.params->rowtot[h]; ++ij){
+        for (long ij = 0; ij < Gab.params->rowtot[h]; ++ij){
             size_t i = Gab.params->roworb[h][ij][0];
             int Gi = Gab.params->psym[i];
             i -= Gab.params->poff[Gi];
@@ -183,7 +183,7 @@ DCFTSolver::compute_unrelaxed_density_OOOO() {
         global_dpd_->buf4_mat_irrep_rd(&Gbb, h);
 
         #pragma omp parallel for
-        for(int ij = 0; ij < Gbb.params->rowtot[h]; ++ij){
+        for (long ij = 0; ij < Gbb.params->rowtot[h]; ++ij){
             size_t i = Gbb.params->roworb[h][ij][0];
             int Gi = Gbb.params->psym[i];
             i -= Gbb.params->poff[Gi];
@@ -824,7 +824,7 @@ DCFTSolver::compute_unrelaxed_density_OVOV() {
         global_dpd_->buf4_mat_irrep_rd(&Gaa, h);
 
         #pragma omp parallel for
-        for(int ia = 0; ia < Gaa.params->rowtot[h]; ++ia){
+        for (long ia = 0; ia < Gaa.params->rowtot[h]; ++ia){
             size_t i = Gaa.params->roworb[h][ia][0];
             int Gi = Gaa.params->psym[i];
             i -= Gaa.params->poff[Gi];
@@ -879,7 +879,7 @@ DCFTSolver::compute_unrelaxed_density_OVOV() {
         global_dpd_->buf4_mat_irrep_rd(&Gab, h);
 
         #pragma omp parallel for
-        for(int ia = 0; ia < Gab.params->rowtot[h]; ++ia){
+        for (long ia = 0; ia < Gab.params->rowtot[h]; ++ia){
             size_t i = Gab.params->roworb[h][ia][0];
             int Gi = Gab.params->psym[i];
             i -= Gab.params->poff[Gi];
@@ -911,7 +911,7 @@ DCFTSolver::compute_unrelaxed_density_OVOV() {
         global_dpd_->buf4_mat_irrep_rd(&Gba, h);
 
         #pragma omp parallel for
-        for(int ia = 0; ia < Gba.params->rowtot[h]; ++ia){
+        for (long ia = 0; ia < Gba.params->rowtot[h]; ++ia){
             size_t i = Gba.params->roworb[h][ia][0];
             int Gi = Gba.params->psym[i];
             i -= Gba.params->poff[Gi];
@@ -979,7 +979,7 @@ DCFTSolver::compute_unrelaxed_density_OVOV() {
         global_dpd_->buf4_mat_irrep_rd(&Gbb, h);
 
         #pragma omp parallel for
-        for(int ia = 0; ia < Gbb.params->rowtot[h]; ++ia){
+        for (long ia = 0; ia < Gbb.params->rowtot[h]; ++ia){
             size_t i = Gbb.params->roworb[h][ia][0];
             int Gi = Gbb.params->psym[i];
             i -= Gbb.params->poff[Gi];
@@ -1060,7 +1060,7 @@ DCFTSolver::compute_unrelaxed_density_VVVV()
         global_dpd_->buf4_mat_irrep_rd(&Gaa, h);
 
         #pragma omp parallel for
-        for(int ab = 0; ab < Gaa.params->rowtot[h]; ++ab){
+        for (long ab = 0; ab < Gaa.params->rowtot[h]; ++ab){
             size_t a = Gaa.params->roworb[h][ab][0];
             int Ga = Gaa.params->psym[a];
             a -= Gaa.params->poff[Ga];
@@ -1094,7 +1094,7 @@ DCFTSolver::compute_unrelaxed_density_VVVV()
         global_dpd_->buf4_mat_irrep_rd(&Gab, h);
 
         #pragma omp parallel for
-        for(int ab = 0; ab < Gab.params->rowtot[h]; ++ab){
+        for (long ab = 0; ab < Gab.params->rowtot[h]; ++ab){
             size_t a = Gab.params->roworb[h][ab][0];
             int Ga = Gab.params->psym[a];
             a -= Gab.params->poff[Ga];
@@ -1126,7 +1126,7 @@ DCFTSolver::compute_unrelaxed_density_VVVV()
         global_dpd_->buf4_mat_irrep_rd(&Gbb, h);
 
         #pragma omp parallel for
-        for(int ab = 0; ab < Gbb.params->rowtot[h]; ++ab){
+        for (long ab = 0; ab < Gbb.params->rowtot[h]; ++ab){
             size_t a = Gbb.params->roworb[h][ab][0];
             int Ga = Gbb.params->psym[a];
             a -= Gbb.params->poff[Ga];
@@ -1212,7 +1212,7 @@ DCFTSolver::compute_relaxed_density_OOOO()
         global_dpd_->buf4_mat_irrep_rd(&Gaa, h);
 
         #pragma omp parallel for
-        for(int ij = 0; ij < Gaa.params->rowtot[h]; ++ij){
+        for (long ij = 0; ij < Gaa.params->rowtot[h]; ++ij){
             size_t i = Gaa.params->roworb[h][ij][0];
             int Gi = Gaa.params->psym[i];
             i -= Gaa.params->poff[Gi];
@@ -1256,7 +1256,7 @@ DCFTSolver::compute_relaxed_density_OOOO()
         global_dpd_->buf4_mat_irrep_rd(&Gab, h);
 
         #pragma omp parallel for
-        for(int ij = 0; ij < Gab.params->rowtot[h]; ++ij){
+        for (long ij = 0; ij < Gab.params->rowtot[h]; ++ij){
             size_t i = Gab.params->roworb[h][ij][0];
             int Gi = Gab.params->psym[i];
             i -= Gab.params->poff[Gi];
@@ -1295,7 +1295,7 @@ DCFTSolver::compute_relaxed_density_OOOO()
         global_dpd_->buf4_mat_irrep_rd(&Gbb, h);
 
         #pragma omp parallel for
-        for(int ij = 0; ij < Gbb.params->rowtot[h]; ++ij){
+        for (long ij = 0; ij < Gbb.params->rowtot[h]; ++ij){
             size_t i = Gbb.params->roworb[h][ij][0];
             int Gi = Gbb.params->psym[i];
             i -= Gbb.params->poff[Gi];
@@ -1476,7 +1476,7 @@ DCFTSolver::compute_relaxed_density_OVOV()
         global_dpd_->buf4_mat_irrep_rd(&Gaa, h);
 
         #pragma omp parallel for
-        for(int ia = 0; ia < Gaa.params->rowtot[h]; ++ia){
+        for (long ia = 0; ia < Gaa.params->rowtot[h]; ++ia){
             size_t i = Gaa.params->roworb[h][ia][0];
             int Gi = Gaa.params->psym[i];
             i -= Gaa.params->poff[Gi];
@@ -1534,7 +1534,7 @@ DCFTSolver::compute_relaxed_density_OVOV()
         global_dpd_->buf4_mat_irrep_rd(&Gab, h);
 
         #pragma omp parallel for
-        for(int ia = 0; ia < Gab.params->rowtot[h]; ++ia){
+        for (long ia = 0; ia < Gab.params->rowtot[h]; ++ia){
             size_t i = Gab.params->roworb[h][ia][0];
             int Gi = Gab.params->psym[i];
             i -= Gab.params->poff[Gi];
@@ -1567,7 +1567,7 @@ DCFTSolver::compute_relaxed_density_OVOV()
         global_dpd_->buf4_mat_irrep_rd(&Gba, h);
 
         #pragma omp parallel for
-        for(int ia = 0; ia < Gba.params->rowtot[h]; ++ia){
+        for (long ia = 0; ia < Gba.params->rowtot[h]; ++ia){
             size_t i = Gba.params->roworb[h][ia][0];
             int Gi = Gba.params->psym[i];
             i -= Gba.params->poff[Gi];
@@ -1668,7 +1668,7 @@ DCFTSolver::compute_relaxed_density_OVOV()
         global_dpd_->buf4_mat_irrep_rd(&Gbb, h);
 
         #pragma omp parallel for
-        for(int ia = 0; ia < Gbb.params->rowtot[h]; ++ia){
+        for (long ia = 0; ia < Gbb.params->rowtot[h]; ++ia){
             size_t i = Gbb.params->roworb[h][ia][0];
             int Gi = Gbb.params->psym[i];
             i -= Gbb.params->poff[Gi];
@@ -1753,7 +1753,7 @@ DCFTSolver::compute_relaxed_density_VVVV()
         global_dpd_->buf4_mat_irrep_rd(&Gaa, h);
 
         #pragma omp parallel for
-        for(int ab = 0; ab < Gaa.params->rowtot[h]; ++ab){
+        for (long ab = 0; ab < Gaa.params->rowtot[h]; ++ab){
             size_t a = Gaa.params->roworb[h][ab][0];
             int Ga = Gaa.params->psym[a];
             a -= Gaa.params->poff[Ga];
@@ -1792,7 +1792,7 @@ DCFTSolver::compute_relaxed_density_VVVV()
         global_dpd_->buf4_mat_irrep_rd(&Gab, h);
 
         #pragma omp parallel for
-        for(int ab = 0; ab < Gab.params->rowtot[h]; ++ab){
+        for (long ab = 0; ab < Gab.params->rowtot[h]; ++ab){
             size_t a = Gab.params->roworb[h][ab][0];
             int Ga = Gab.params->psym[a];
             a -= Gab.params->poff[Ga];
@@ -1827,7 +1827,7 @@ DCFTSolver::compute_relaxed_density_VVVV()
         global_dpd_->buf4_mat_irrep_rd(&Gbb, h);
 
         #pragma omp parallel for
-        for(int ab = 0; ab < Gbb.params->rowtot[h]; ++ab){
+        for (long ab = 0; ab < Gbb.params->rowtot[h]; ++ab){
             size_t a = Gbb.params->roworb[h][ab][0];
             int Ga = Gbb.params->psym[a];
             a -= Gbb.params->poff[Ga];
